chefão e passagem entre fases adicionados

+++ headers/Inimigo1.h
@@ -4,7 +4,7 @@
 class Inimigo1 : public Inimigo
 {
 private:
-	float posxInicial;
+	int posxInicial;
 	bool teste;
 public:
 	Inimigo1(Vetor2F pos = (0.0, 0.0), Vetor2F vel = (0.0, 0.0), Vetor2F tam = (0.0,0.0));
